+++ internal/config/systemd_credential_other.go
@@ -1,0 +1,9 @@
+//go:build !linux
+
+package config
+
+import "os"
+
+func isSystemdCredential(_ string, _ os.FileInfo) bool {
+	return false
+}